+++ dawn/test/unit-test/dawn/Support/CMakeLists.txt
@@ -24,7 +24,7 @@ add_executable(${executable}
   TestRemoveIf.cpp
   TestRangeToString.cpp
   TestType.cpp
-  TestIteratorAdaptors.cpp
+  TestIteratorAdapters.cpp
 )
 target_link_libraries(${executable} DawnSupport DawnUnittest gtest)
 target_add_dawn_standard_props(${executable})
